requirements.txt
# Django & Web Framework
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0

# Monitoring & Observability
django-prometheus==2.3.1
prometheus-client==0.19.0
psutil==5.9.8

# Database
psycopg2-binary==2.9.9

# WSGI Server
gunicorn==21.2.0

# Development
python-decouple==3.8
requests==2.31.0
